More refactoring + adjusting unit tests
