+++ ironsourceatomsdk/src/test/java/io/ironsourceatom/sdk/ReportHandlerIntegrationTest.java
@@ -30,12 +30,13 @@ public class ReportHandlerIntegrationTest {
     @Before
     public void reset() {
         mClient.mBackedMock.clear();
+        mReportService.init(RuntimeEnvironment.application);
     }
 
     @Test
     public void testPostSuccess() throws Exception {
-        mHandler.handleReport(newReport(SdkEvent.POST_SYNC, event1));
-        mHandler.handleReport(newReport(SdkEvent.POST_SYNC, event2));
+        mReportService.handleReport(newReport(SdkEvent.POST_SYNC, event1));
+        mReportService.handleReport(newReport(SdkEvent.POST_SYNC, event2));
         assertEquals(mClient.get(TABLE1), new JSONArray("[{" +
                 "\"data\":\"ib-data\"," +
                 "\"table\":\"ib_test\"," +
@@ -53,8 +54,8 @@ public class ReportHandlerIntegrationTest {
     @Test
     public void testPostFailed() {
         mClient.setNext(503);
-        mHandler.handleReport(newReport(SdkEvent.POST_SYNC, event1));
-        mHandler.handleReport(newReport(SdkEvent.POST_SYNC, event2));
+        mReportService.handleReport(newReport(SdkEvent.POST_SYNC, event1));
+        mReportService.handleReport(newReport(SdkEvent.POST_SYNC, event2));
         assertEquals(mAdapter.count(null), 2);
         assertEquals(mAdapter.getTables().size(), 2);
     }
@@ -63,7 +64,7 @@ public class ReportHandlerIntegrationTest {
     public void testTrackEvent() {
         mConfig.setBulkSize(Integer.MAX_VALUE);
         for (int i = 1; i <= 10; i++) {
-            mHandler.handleReport(newReport(SdkEvent.ENQUEUE, event1));
+            mReportService.handleReport(newReport(SdkEvent.ENQUEUE, event1));
             assertEquals(mAdapter.count(null), i);
         }
         assertEquals(mAdapter.getTables().size(), 1);
@@ -71,7 +72,7 @@ public class ReportHandlerIntegrationTest {
 
     @Test
     public void testTrackError() throws JSONException {
-        mHandler.handleReport(newReport(SdkEvent.REPORT_ERROR, event1));
+        mReportService.handleReport(newReport(SdkEvent.REPORT_ERROR, event1));
         assertEquals(mClient.get(TABLE1), new JSONArray("[{" +
                 "\"data\":\"ib-data\"," +
                 "\"table\":\"ib_test\"," +
@@ -86,7 +87,7 @@ public class ReportHandlerIntegrationTest {
         for (int i = 1; i <= 10; i++) {
             final Map<String, String> event = new HashMap<>(event1);
             event.put(ReportData.DATA, String.valueOf(i));
-            mHandler.handleReport(newReport(SdkEvent.ENQUEUE, event));
+            mReportService.handleReport(newReport(SdkEvent.ENQUEUE, event));
         }
         assertEquals(mAdapter.count(null), 0);
         assertEquals(mAdapter.getTables().size(), 0);
@@ -124,9 +125,9 @@ public class ReportHandlerIntegrationTest {
         for (int i = 1; i <= 10; i++) {
             final Map<String, String> event = new HashMap<>(event1);
             event.put(ReportData.DATA, String.valueOf(i));
-            mHandler.handleReport(newReport(SdkEvent.ENQUEUE, event));
+            mReportService.handleReport(newReport(SdkEvent.ENQUEUE, event));
             event.put(ReportData.TABLE, TABLE2);
-            mHandler.handleReport(newReport(SdkEvent.ENQUEUE, event));
+            mReportService.handleReport(newReport(SdkEvent.ENQUEUE, event));
         }
         assertEquals(mAdapter.count(null), 0);
         assertEquals(mAdapter.getTables().size(), 0);
@@ -137,21 +138,21 @@ public class ReportHandlerIntegrationTest {
     // Events to test
     final String TABLE1 = "ib_test", TOKEN1 = "ib_token", DATA1 = "ib-data";
     final String TABLE2 = "ic_test", TOKEN2 = "ic_token", DATA2 = "ic-data";
-    final Map<String, String>   event1   = new HashMap<String, String>() {{
+    final Map<String, String>   event1         = new HashMap<String, String>() {{
         put(ReportData.DATA, DATA1);
         put(ReportData.TOKEN, TOKEN1);
         put(ReportData.TABLE, TABLE1);
     }};
-    final Map<String, String>   event2   = new HashMap<String, String>() {{
+    final Map<String, String>   event2         = new HashMap<String, String>() {{
         put(ReportData.DATA, DATA2);
         put(ReportData.TOKEN, TOKEN2);
         put(ReportData.TABLE, TABLE2);
     }};
     // MockBackend
-    final TestsUtils.MockPoster mClient  = new TestsUtils.MockPoster();
-    final IsaConfig             mConfig  = IsaConfig.getInstance(RuntimeEnvironment.application);
-    final StorageApi            mAdapter = new DbAdapter(RuntimeEnvironment.application);
-    final ReportHandler         mHandler = new ReportHandler(RuntimeEnvironment.application) {
+    final TestsUtils.MockPoster mClient        = new TestsUtils.MockPoster();
+    final IsaConfig             mConfig        = IsaConfig.getInstance(RuntimeEnvironment.application);
+    final StorageApi            mAdapter       = new DbAdapter(RuntimeEnvironment.application);
+    final ReportService         mReportService = new ReportService() {
         @Override
         protected StorageApi getStorage(Context context) {
             return mAdapter;
